fix(PaymentMethodRequest.java): fix totalAmount data type to Long

+++ src/main/java/it/gov/pagopa/afm/calculator/model/paymentmethods/PaymentMethodRequest.java
@@ -26,7 +26,7 @@ public class PaymentMethodRequest {
     UserDevice userDevice;
 
     @NotNull
-    Integer totalAmount;
+    Long totalAmount;
 
     @NotNull
     @NotEmpty
